feat: 统一经纬度值类型

+++ database/migrations/2023_04_10_104850_create_scenic_spots_table.php
@@ -20,8 +20,8 @@ class CreateScenicSpotsTable extends Migration
             $table->integer('category_id')->comment('景区分类id');
             $table->string('name')->comment('景区名称');
             $table->string('level')->default('')->comment('景区等级');
-            $table->float('longitude')->comment('经度');
-            $table->float('latitude')->comment('纬度');
+            $table->decimal('longitude', 9, 6)->comment('经度');
+            $table->decimal('latitude', 8, 6)->comment('纬度');
             $table->string('address')->comment('具体地址');
             $table->float('rate')->default(0)->comment('景区评分');
             $table->string('video')->default('')->comment('视频');
